requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyascii==0.3.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.11.2
certifi==2023.11.17
charset-normalizer==3.3.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.1
django-autoslug==1.9.9
django-filter==23.5
django-import-export==3.3.7
django-modelcluster==6.2.1
django-otp==1.3.0
django-permissionedforms==0.1
django-phonenumber-field==7.3.0
django-redis==5.4.0
django-taggit==4.0.0
django-treebeard==4.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filetype==1.2.0
frozenlist==1.4.1
gunicorn==21.2.0
html5lib==1.1
idna==3.6
l18n==2021.3
laces==0.1.1
MarkupPy==1.14
multidict==6.0.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
phonenumbers==8.13.28
pillow==10.2.0
pillow-heif==0.14.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytils==0.4.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.2
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tablib==3.5.0
telepath==0.3.1
tzdata==2023.4
urllib3==2.1.0
wagtail==6.0.1
wagtail-import-export-tool==0.9.5
webencodings==0.5.1
Willow==1.6.3
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
